+++ msl/equipment/resources/greisinger/__init__.py
@@ -1,0 +1,4 @@
+"""
+Resources for equipment from `Greisinger <https://www.greisinger.de/>`_.
+"""
+from .gmh3000 import GMH3000