+++ CHANGES.rst
@@ -5,6 +5,14 @@ Release Notes
 Version 0.2.0 (in development)
 ==============================
 
+* Added
+
+  - The :class:`~msl.equipment.resources.greisinger.gmh3000.GMH3000` resource.
+
+* Removed
+
+  - Support Python 2.7, 3.5, 3.6 and 3.7
+
 Version 0.1.0 (2023-06-18)
 ==========================
 Initial release.